refactor(core): extract data management to ChartDataManager
Move data/fields/methods from Chart god class into ChartDataManager
with dependency injection (no Chart instance reference).

Migrated: _internalData, _dataFetcher, _dataBuffer, comparison
buffers, signals (_dataSignal, _symbolsSignal), incremental load
hint, lastVisibleRange, scrollToRight, setSymbols, setData, and
all related helpers.

Chart retains thin delegation layer (63 lines data API) and rendering
pipeline accesses data via dataManager.getInternalData() etc.
